+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Added a system command for reversing the current sort order.
   ([#46](https://github.com/davep/peplum/pull/46))
 - Added `--theme` as a command line switch.
+- Added `--sort-by` as a command line switch.
 - Added support for taking a PEP number on the command line and jumping to
   it on startup.